+++ adapters/LiftoffMonetize/LiftoffMonetizeAdapter/GADMAdapterVungleUtils.h
@@ -22,8 +22,8 @@
 NSError *_Nonnull GADMAdapterVungleErrorWithCodeAndDescription(GADMAdapterVungleErrorCode code,
                                                                NSString *_Nonnull description);
 
-/// Returns a NSError converted from the error object from Liftoff Monetize's SDK into the AdMob error
-/// format The localized description will contain the Liftoff Monetize error code and the
+/// Returns a NSError converted from the error object from Vungle SDK into the AdMob
+/// error format The localized description will contain the Liftoff Monetize error code and the
 /// description from the Vungle SDK.
 NSError *_Nonnull GADMAdapterVungleErrorToGADError(GADMAdapterVungleErrorCode code,
                                                    NSInteger vungleCode,
@@ -38,11 +38,14 @@ NSError *_Nonnull GADMAdapterVungleInvalidPlacementErrorWithCodeAndDescription(v
 /// NSLocalizedFailureReasonErrorKey values set to |description|, specifically for invalid app id.
 NSError *_Nonnull GADMAdapterVungleInvalidAppIdErrorWithCodeAndDescription(void);
 
-/// Returns a GADAdSize object that is valid for Liftoff Monetize's SDK
+/// Returns a GADAdSize object that is valid for Vungle SDK.
 GADAdSize GADMAdapterVungleAdSizeForAdSize(GADAdSize adSize);
 
-/// Returns a Liftoff Monetize BannerSize from the provided GADAdSize
+/// Returns a Liftoff Monetize BannerSize from the provided GADAdSize.
 BannerSize GADMAdapterVungleConvertGADAdSizeToBannerSize(GADAdSize adSize);
+
+/// Safely adds |object| to |set| if |object| is not nil.
+void GADMAdapterVungleMutableSetAddObject(NSMutableSet *_Nullable set, NSObject *_Nonnull object);
 
 @interface GADMAdapterVungleUtils : NSObject