+++ adapters/LiftoffMonetize/LiftoffMonetizeAdapter/GADMediationAdapterVungle.h
@@ -13,21 +13,21 @@
 // limitations under the License.
 
 #import <Foundation/Foundation.h>
-@import GoogleMobileAds;
+#import <GoogleMobileAds/GoogleMobileAds.h>
 
 typedef NS_ENUM(NSInteger, GADMAdapterVungleErrorCode) {
   /// Missing server parameters.
   GADMAdapterVungleErrorInvalidServerParameters = 101,
-  /// An ad is already loaded for this network configuration. Liftoff Monetize's SDK cannot load
+  /// An ad is already loaded for this network configuration. Vungle SDK cannot load
   /// a second ad for the same placement ID.
   GADMAdapterVungleErrorAdAlreadyLoaded = 102,
   /// Banner Size Mismatch.
   GADMAdapterVungleErrorBannerSizeMismatch = 103,
-  /// Liftoff Monetize's SDK could not render the banner ad.
+  /// Vungle SDK could not render the banner ad.
   GADMAdapterVungleErrorRenderBannerAd = 104,
-  /// Liftoff Monetize's SDK only supports loading 1 banner ad at a time, regardless of placement ID.
+  /// Vungle SDK only supports loading 1 banner ad at a time, regardless of placement ID.
   GADMAdapterVungleErrorMultipleBanners = 105,
-  /// Liftoff Monetize's SDK sent a callback saying the ad is not playable.
+  /// Vungle SDK sent a callback saying the ad is not playable.
   GADMAdapterVungleErrorAdNotPlayable = 106
 };