+++ adapters/LiftoffMonetize/CHANGELOG.md
@@ -1,10 +1,15 @@
 ## Liftoff Monetize iOS Mediation Adapter Changelog
 
-#### Next Version
+#### Version 7.0.0 (In Progress)
 - Rebranded adapter name from Vungle to Liftoff Monetize.
+- Verified compatibility with Vungle SDK 7.0.0.
 - Removed support of the `armv7` architecture.
 - Now requires minimum iOS version 11.0.
 - Now requires Google Mobile Ads SDK version 10.4.0 or higher.
+
+Built and tested with:
+- Google Mobile Ads SDK version 10.4.0.
+- Vungle SDK version 7.0.0.
 
 #### [Version 6.12.3.0](https://dl.google.com/googleadmobadssdk/mediation/ios/vungle/VungleAdapter-6.12.3.0.zip)
 - Verified compatibility with Vungle SDK 6.12.3.